feat[transport]: Explicit length fields

+++ SSH-Core/src/main/java/de/rub/nds/sshattacker/core/constants/DisconnectReason.java
@@ -18,25 +18,25 @@ public enum DisconnectReason {
      *  - https://www.iana.org/assignments/ssh-parameters/ssh-parameters.xhtml#ssh-parameters-3
      */
     // [ RFC 4253 ]
-    SSH_DISCONNECT_HOST_NOT_ALLOWED_TO_CONNECT((byte) 1),
-    SSH_DISCONNECT_PROTOCOL_ERROR((byte) 2),
-    SSH_DISCONNECT_KEY_EXCHANGE_FAILED((byte) 3),
-    SSH_DISCONNECT_RESERVED((byte) 4),
-    SSH_DISCONNECT_MAC_ERROR((byte) 5),
-    SSH_DISCONNECT_COMPRESSION_ERROR((byte) 6),
-    SSH_DISCONNECT_SERVICE_NOT_AVAILABLE((byte) 7),
-    SSH_DISCONNECT_PROTOCOL_VERSION_NOT_SUPPORTED((byte) 8),
-    SSH_DISCONNECT_HOST_KEY_NOT_VERIFIABLE((byte) 9),
-    SSH_DISCONNECT_CONNECTION_LOST((byte) 10),
-    SSH_DISCONNECT_BY_APPLICATION((byte) 11),
-    SSH_DISCONNECT_TOO_MANY_CONNECTIONS((byte) 12),
-    SSH_DISCONNECT_AUTH_CANCELLED_BY_USER((byte) 13),
-    SSH_DISCONNECT_NO_MORE_AUTH_METHODS_AVAILABLE((byte) 14),
-    SSH_DISCONNECT_ILLEGAL_USER_NAME((byte) 15);
+    SSH_DISCONNECT_HOST_NOT_ALLOWED_TO_CONNECT(1),
+    SSH_DISCONNECT_PROTOCOL_ERROR(2),
+    SSH_DISCONNECT_KEY_EXCHANGE_FAILED(3),
+    SSH_DISCONNECT_RESERVED(4),
+    SSH_DISCONNECT_MAC_ERROR(5),
+    SSH_DISCONNECT_COMPRESSION_ERROR(6),
+    SSH_DISCONNECT_SERVICE_NOT_AVAILABLE(7),
+    SSH_DISCONNECT_PROTOCOL_VERSION_NOT_SUPPORTED(8),
+    SSH_DISCONNECT_HOST_KEY_NOT_VERIFIABLE(9),
+    SSH_DISCONNECT_CONNECTION_LOST(10),
+    SSH_DISCONNECT_BY_APPLICATION(11),
+    SSH_DISCONNECT_TOO_MANY_CONNECTIONS(12),
+    SSH_DISCONNECT_AUTH_CANCELLED_BY_USER(13),
+    SSH_DISCONNECT_NO_MORE_AUTH_METHODS_AVAILABLE(14),
+    SSH_DISCONNECT_ILLEGAL_USER_NAME(15);
 
-    public final byte id;
+    public final int id;
 
-    public static final Map<Byte, DisconnectReason> map;
+    public static final Map<Integer, DisconnectReason> map;
 
     static {
         map = new TreeMap<>();
@@ -45,15 +45,15 @@ public enum DisconnectReason {
         }
     }
 
-    DisconnectReason(byte id) {
+    DisconnectReason(int id) {
         this.id = id;
     }
 
-    public static String getNameByID(byte id) {
+    public static String getNameByID(int id) {
         return map.get(id).toString();
     }
 
-    public static DisconnectReason fromId(byte id) {
+    public static DisconnectReason fromId(int id) {
         return map.get(id);
     }
 }
